very aggressive highscore cache
this makes the user page and allscores page load very fast after
previously already loading it a couple minutes before. this probably
starts to suck as the site gets larger. a huge amount of scores should
fit in memory but the initial load might be really bad. will have to
look at later when it gets really bad